change scripts structure
